server/requirements.txt
asgiref==3.5.2
Django==4.1
django-cors-headers==3.13.0
django-js-asset==2.0.0
django-mptt==0.13.4
djangorestframework==3.13.1
Pillow==9.2.0
psycopg2==2.9.3
pytz==2022.1
sqlparse==0.4.2
